ktra cart item ton tai + chinh chu khi add, delete, update

+++ eComSolution.Service/Catalog/Carts/CartService.cs
@@ -18,7 +18,17 @@ namespace eComSolution.Service.Catalog.Carts
             _context = context;
         }
         public async Task<ApiResult<int>> AddToCart(int userId, AddToCartRequest request)
-        {           
+        {    
+            // check valid properties request
+            if(request.IsValid()==false)
+                return new ApiResult<int>(false, Message:"Thông tin không hợp lệ, vui lòng nhập lại");  
+            // check product detail
+            var product_detail = await _context.ProductDetails.FirstOrDefaultAsync(x=>x.Id==request.ProductDetail_Id);     
+            if(product_detail==null)
+                return new ApiResult<int>(false, Message:"Phân loại hàng không tồn tại"); 
+            if(product_detail.IsDeleted==true)
+                return new ApiResult<int>(false, Message:"Phân loại hàng đã bị xóa"); 
+                
             int stock = (await _context.ProductDetails
                     .Where(x=>x.Id == request.ProductDetail_Id)
                     .FirstOrDefaultAsync()).Stock; 
@@ -120,20 +130,30 @@ namespace eComSolution.Service.Catalog.Carts
             return new ApiResult<List<CartItem>>(true, ResultObj:data);
         }
 
-        public async Task<ApiResult<int>> RemoveFromCart(int cartId)
+        public async Task<ApiResult<int>> RemoveFromCart(int userId, int cartId)
         {
             var cart_item = await _context.Carts.Where(x=>x.Id == cartId).FirstOrDefaultAsync();
+            if(cart_item==null)
+                return new ApiResult<int>(false, Message:$"Không tồn tại cart item với Id:{cartId}"); 
+            if(cart_item.UserId!=userId)
+                return new ApiResult<int>(false, Message:$"Bạn không có cart item với Id: {cartId}"); 
+
             _context.Carts.Remove(cart_item);
             await _context.SaveChangesAsync();
             return new ApiResult<int>(true, Message:"Bỏ sản phẩm khỏi giỏ thành công!");
         }
 
-        public async Task<ApiResult<int>> RemoveMultiCarts(List<int> cartIds)
+        public async Task<ApiResult<int>> RemoveMultiCarts(int userId, List<int> cartIds)
         {
 
-            foreach(var id in cartIds)
+            foreach(var cartId in cartIds)
             {
-                var cart_item = await _context.Carts.Where(x=>x.Id == id).FirstOrDefaultAsync();
+                var cart_item = await _context.Carts.Where(x=>x.Id == cartId).FirstOrDefaultAsync();
+                if(cart_item==null)
+                    return new ApiResult<int>(false, Message:$"Không tồn tại cart item với Id:{cartId}"); 
+                if(cart_item.UserId!=userId)
+                    return new ApiResult<int>(false, Message:$"Bạn không có cart item với Id: {cartId}");
+
                 _context.Carts.Remove(cart_item);          
             }
     
@@ -141,17 +161,22 @@ namespace eComSolution.Service.Catalog.Carts
             return new ApiResult<int>(true, Message:"Bỏ sản phẩm khỏi giỏ thành công!");
         }
 
-        public async Task<ApiResult<int>> UpdateCartItem(UpdateCartItemRequest request)
+        public async Task<ApiResult<int>> UpdateCartItem(int userId, UpdateCartItemRequest request)
         {
+            // check valid properties request
+            if(request.IsValid()==false)
+                return new ApiResult<int>(false, Message:"Thông tin không hợp lệ, vui lòng nhập lại"); 
             var cart_item = await _context.Carts.Where(x=>x.Id==request.CartId).FirstOrDefaultAsync();
             if(cart_item==null)
-                return new ApiResult<int>(false, Message:"Giỏ hàng không tồn tại!"); 
+                return new ApiResult<int>(false, Message:$"Không tồn tại cart item với Id:{request.CartId}"); 
+            if(cart_item.UserId!=userId)
+                return new ApiResult<int>(false, Message:$"Bạn không có cart item với Id: {request.CartId}"); 
 
             int stock = (await _context.ProductDetails
                     .Where(x=>x.Id == cart_item.ProductDetail_Id)
                     .FirstOrDefaultAsync()).Stock; 
 
-            if(request.Quantity > stock )
+            if(request.Quantity > stock)
                 return new ApiResult<int>(false, Message:"Quá số lượng tồn kho!");
 
             cart_item.Quantity = request.Quantity;
